Various improvements.
* ShapeEditing uses double to allow subpixel positioning.
* Factored MouseSelectionBrain out of SelectionBrain.
* SelectionBrain responds to selected notifications; sets
  managed list members isSelected flag.
* Changed default shape creation from Prepend to Append.
  Created shapes will be logically on the bottom of the stack.

+++ draw/cross_shape.h
@@ -46,7 +46,7 @@ struct CrossShapeTemplates: public ShapeCommon<CrossGroup, CrossView>
 
         std::unique_ptr<Drag> ProcessMouseDown(
             std::shared_ptr<ShapeControl> control,
-            const tau::Point2d<int> &click,
+            const tau::Point2d<double> &click,
             const wxpex::Modifier &modifier,
             CursorControl cursor) override
         {
@@ -75,7 +75,7 @@ struct CreateCross
 {
     std::optional<ShapeValue> operator()(
         const Drag &,
-        const tau::Point2d<int> position)
+        const tau::Point2d<double> position)
     {
         auto cross = Cross{};
         cross.center = position;
