+++ draw/quad_brain.h
@@ -44,7 +44,7 @@ public:
 
 protected:
     std::shared_ptr<Shape> MakeShape_(
-        const tau::Point2d<int> &end) const override
+        const tau::Point2d<double> &end) const override
     {
         if (end == this->start_)
         {
@@ -143,7 +143,7 @@ public:
 
 protected:
     std::shared_ptr<Shape> MakeShape_(
-        const tau::Point2d<int> &end) const override
+        const tau::Point2d<double> &end) const override
     {
         if (end == this->start_)
         {
@@ -170,7 +170,7 @@ public:
 
     DragQuadLine(
         size_t index,
-        const tau::Point2d<int> &start,
+        const tau::Point2d<double> &start,
         const QuadLines &quadLines,
         std::shared_ptr<ShapeControl> control,
         const DerivedShape &startingShape)
@@ -188,7 +188,7 @@ public:
 
 protected:
     std::shared_ptr<Shape> MakeShape_(
-        const tau::Point2d<int> &end) const override
+        const tau::Point2d<double> &end) const override
     {
         const auto &line = this->lines_[this->index_];
         double lineAngle = line.GetAngleDegrees();
